IO validation added

+++ seminar-1/src/plus.cpp
@@ -1,5 +1,5 @@
 #include "plus.h"
 
-LL sum(LL first_number, LL second_number) {
+long long sum(long long first_number, long long second_number) {
     return first_number + second_number;
 }
